requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
PyYAML==3.12
requests==2.18.4
sender==0.3
urllib3==1.22
